cleaned up some of the code
