+++ views/index.handlebars
@@ -18,9 +18,9 @@
     }
 </style>
 <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-    <a class="navbar-brand" href="#">
+    <div class="navbar-brand">
         CriShipped
-    </a>
+    </div>
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavDropdown"
         aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
@@ -35,6 +35,9 @@
             </li>
             <li class="nav-item">
                 <a class="nav-link" href="#">Sell Supplies</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="#">Discussion Forum</a>
             </li>
         </ul>
     </div>
